+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 
 ## [Unreleased]
 
+### Added
+
+- Option to cancel batch TEI-transcription requests in progress and pending.
+
 ### Fixed
 
 - Improve TEI encoding prompts.